Add the payment gateway logic and related specs
We add the pos_order_id field to an order, since we can't search for the
order_number with the Spotify API we need to be able to find it by the
ID. Also, I didn't want to use the number field of an order since I feel
like the pos_order_number should belong there.

In this commit, we include the related spree_gateway and active_merchant
gem in order to allow the new shopify gateway to work. Then there is a
single spec (more to come) that make sure that we can successfully
refund an order that was made in Shopify from the Solidus perspective by
using the proper Gateway.