Basic scaffolding to allow web hooks reception
This also includes how I was thinking to test the web hooks. There would
be a Shopify request object that would include methods that will fetch
"test" data as sent by Shopify. That way we could easily modify that
test file and pass that as the request body, exactly how Shopify would
do it.

https://github.com/glossier/glossier-v2/commit/27d62a2bfd4dacd420d9d77b817f1ce0bc02bef8