WIP - Spec writing for the order importer
There is still a lot to spec out. Most of the current specs are red due
to the fact that an order still do not complete.

There was also refactoring here and there in order to support what I
have in mind.

I have decided to go with Webmock instead of VCR since I feel like VCR
is too dirty for what we want to achieve. We want to be able to easily
modify the text file in order to change the content of the txt file with
different value if we ever want to.