WIP - first draft for the image import
There are tons of problems coming up. Mainly, we have a weird way of
setting and fetching the images on Glossier, I am still unsure how we'll
go around that in order to find the proper images. Right now I am using
placeholder images.

Also, in order to assign an image to a variant, the variant has to be
already saved. So we need to save the product, that saves the variant
then we assign the images to the product and when we save the product
again, it will assign the image to the variant. It is a really weird
process but I haven't found a way around that... Yet!