Make sure to re-create product/variant if not exist anymore
If we were to create a product on Solidus and sync it with Shopify then
delete the product from Shopify, the next time we would try to sync the
product/variant it would just crash.

Shopify would return an exception and would halt the process, in that
situation we catch the error and we re-create it to make sure that the
product exists in Shopify.