Add more specs to make sure that everything works
Decided to temporarily remove VCR because it caused a lot of problems.
Since factory girl creates a master variant with a random sku and all
the other variants with random skus, it's hard to sync of with VCR.

I find it insane to mock each skus individually, there must be a better
way.