Add the stock exporter class and task
That task has to be set to be run at a certain interval. This will
update the available quantity of a variant with the available quantity
that was set on Solidus.

This has to be done at a set interval, else if we run the exporter each
time the stock changes, there could be some concurrent requests that get
runs at different time and propagate wrong results to shopify.