Make sure that the stock exporter correctly works with specs
There was a couple of field missing in order for Shopify to register the
stock quantity. In fact, `inventory_management` had to be set to
`shopify`.

Also, `old_inventory_quantity` has to be set while setting
`inventory_quantity` in order for Shopify to correctly register the
change.

Also added more specs in order to improve the coverage of your spec
suite.