Preliminary Cleanup of Gemfile
It's strongly preferable to keep dependencies in the gemspec, rather than the
Gemfile, particularly for base dependencies. This allows RubyGems/Bundler to
resolve the dependency tree. The exception is, for obvious reasons, gems not
hosted on RubyGems.

I also removed a development dependency on 'braintree', on which the test suite
doesn't appear to depend. (Maybe it did at one point?)