Allow the passing of a HTML renderer for the product converter
Instead of having our own method that makes it works only under a
certain scenario, we have opted to go with Redcarpet and convert the
Markdown content to HTML content. That way the content will be displayed
better than our previous ways of doing things.

Also, this let the user change the renderer to whatever rows their boat.

* This also moves the gems from the gemfile to the gemspec file.