Support for Voiding Shopify Transactions
This commit ended up getting a bit out of hand. The original aim was to
obsolesce the Voider service object. Unfortunately, this revealed the absence
of a request spec for voiding a transaction on Shopify. This request spec, once
written, revealed issues with the implementation.

After some wrangling, the voiding workflow is implemented and covered with
tests. Some refactoring remains, but nothing worth delaying a commit over.