Separate updaters and exporters
The way that it works right now is that the exporter will export the
whole product, which means all of its variant.

Updaters will now only update self. So if we go in the admin and we
decide to update a unique product or a unique variant, then it will
update itself instead of saving the whole product will all of its
variant.

There is still some work to do in order to reduce the code that is
duplicated but that should still be good for CRing.

The goal is to run the exporter only once at the beginning of the
project, so fill the Shopify store with proper data. We were kind of
obligated to do that, since if we don't assign any variant to the
product when we create it, it assigns a dummy variant that is hard to
get rid of. I tried just to export the master variant when creating a
new product but it was really pain in the arse. This came out to be the
best solution I could come up with.