Add the logic to export the stock to Shopify
I have detected a problem with the base_decorator where it would not
find the proper variant. I was always testing with a Shopify Shop that
was empty and now that we have some real data a problem came out.

We were always returning all the variants even though we were trying to
find one with a specific id and we first picking the first one from that
array.

Instead, I've decided to use the `find` method that they are already
using in their ShopifyAPI gem and catching the 404 error. When that
exception is caught we create a new instance of that object and return
it.

This also regenerate the cassettes, due to the fact that the
base_decorator has changed which indirectly make the requests change.