Update the shopify id field without triggering the update callback
Problematic case:

Once a Solidus gets updated, we want to trigger an after_update event
that will trigger the task that will update this product to Shopify.

When we update it on Shopify, in a certain case the product on Shopify
may have been deleted so we need to re-create it. When we re-create it,
we need to save the shopify_product_id to the Solidus product. And there
goes an infinite loop

This solves that.