Add a "do not export to shopify" flag on a product
There are some scenarios where we may not want a solidus product to be
updated onto Solidus. Having that attr_accessor flag will let us answer
those cases.

Also, this is really useful to us in the sense that we do not want the
product to be already exported on Shopify when we want to test the
product exporter. In order to run our spec, we are creating solidus
object and previously when we would create those object, they would
already be exported to Shopify.