Save the Solidus variant on Shopify on callbacks
With setting a flag like that, this causes many problems. One of them is
that during associating parts to bundles, we can't set the flag to say
that the product should not be exported to Shopify does not work and it
always exports it. Which is bad. We need to find a work around that

This is still WIP, since the bundle specs are not working anymore, due
to the association saving the product and sending it to Shopify.