Better express the dependency on spree_product_asssembly
A major problem right now is that the way we are using the assembly gem
in our fork is different of the original implementation.

Instead of a product having parts (which is the case in our fork),
variants have parts (in the original implementation). We'll have to
figure something out.

This solves https://github.com/glossier/solidus_retail/issues/37