Make sure the order is marked as shipped
We were marking the shipment as shipped before we would complete it.
During the completion process it was resetting the shipment status and
put it back to pending.

Approving it solved our issue, this is the host app flow and I would
image that this is the normal flow too.

https://app.clubhouse.io/dynamomtl/story/100/when-a-customer-creates-an-order-on-shopify-i-expect-the-imported-order-on-solidus-to-be-mark-as-shipped