Add logic to not re-send refund to Shopify
The problem I am facing is that we wanted to refund orders on Shopify
when they were refunded on Solidus. Which works fine at the moment.

But now that we want it the other way around, which is saving the refund
on Solidus when it's done in Solidus, this gets a little bit more crazy.

We will receive the event and we will proceed to do the Solidus refund
when a Shopify refund is done. When creating a Solidus refund at a later
step it tries to do a refund on the Gateway, but in our scenario the
Gateway is Shopify.

Which makes the following scenario happen: Shopify Refund -> Solidus
Refund -> ShopifyRefund and then it crashes because there are nothing
more to refund on Shopify.

This is a WIP solution for this problem.