Use the presenter to abstract some of the logic
It's not the goal of the converter to know how to have the logic to know
what's title of the variant. The converter should only have one goal,
which is to transform the variant into a hash.

I've moved the logic into the presentation layer, of which I had
forgotten.