Remove the expected headers from request stubbing
There was an issue where the headers that were expected in order for the
requests to be stubbed were invalid. I've decided to remove the instead
of putting the right headers, since we don't totally care about what
headers the request must have in order for it to be mocked.

We want it to be more generic and always mock the request based on the
URL/VERB of the request and not also based on the HEADERS. At least for
now!