Add the proper tax adjustments on refund
It seems like that when you create a refund on Solidus, it removes all
the adjustments associated with an order.

The adjustments is what contained the taxes, so by removing the
adjustments, the order decreases prices and thus we refund the client
with a lower amount of money than they actually paid.