Add reimbursement tax calculator
The Shopify taxes are not imported has taxes in Solidus. We create
an adjustment for every tax type. When we need to calculate the tax
for the refund process, we need to retrieve those adjustments and
count them has taxes so we can refund the proper amount to the client.