Add a custom refund calculator
This custom refund calculator will take care of the bundled products.

The previous default refund calculator would use the line_item price to
calculate the return_item price. The way that bundle works, is that a
line_item contains multiple inventory units.

So if a bundle product line_item was set to 65$, each inventory unit
being returned that were part of the bundled product would be priced at
65$.

We can't simply device the number of inventory units associated with
bundled line_item because some items have more value than the others. We
try to calculate a certain ratio so the item we reimburse is fair.