Update Namespace On Classes (#83)
* Namespace Updater Classes

The Retail gem has some pretty aggressive namespacing going on. Due to
this (I believe), that when updating bundled products on Shopify, the
BundledProductExporter won't be firing within the Retail namespace. This
will cause the update to fail and halt the entire request/response
lifecycle.

* Rubocop Cleanup.

Travis runs Rubocop and will fail if anything is wrong. Remove Rubocop
from running on bin files and clean up several of the current failures.