Updates To Bundled Products. (#84)
* Bundled Product Uploads

The way bundled products works now has changed. To upload parts we need
to look at the first variant of the product and not the product's master
variant. This augments this pointer and greens up the affected tests.

* New Product Bundles Can Be Downloaded From Retail

Augment the GeneratePosOrder class to be able to consume bundled
products that are assigned to the first variant and not the master
variant.